Refactoring of names and cleanup of unused files.
