+++ client/src/components/RoadMeasurements/RideCards.tsx
@@ -1,90 +1,75 @@
-import {FC, ReactNode, useEffect, useState} from "react";
-import {List, ListRowRenderer} from "react-virtualized";
+import { FC, ReactNode, useEffect } from 'react';
+import { List, ListRowRenderer } from 'react-virtualized';
 
 import Checkbox from '../Checkbox';
 
-import {RideMeta, TripsOptions} from '../../models/models'
+import { RideMeta } from '../../models/models';
 
-import {useMetasCtx} from "../../context/MetasContext";
-import OptionsSelector from "./OptionsSelector";
+import { useMetasCtx } from '../../context/MetasContext';
 
 
 interface CardsProps {
-    showMetas: SelectMeta[]
-    onClick: (meta: SelectMeta, i: number, isChecked: boolean) => void;
+	showMetas: SelectMeta[];
+	onClick: (meta: SelectMeta, i: number, isChecked: boolean) => void;
+	height: number;
 }
 
-const Cards: FC<CardsProps> = ({showMetas, onClick}) => {
-    const renderRow: ListRowRenderer = ({index, key, style}): ReactNode => {
-        const meta = showMetas[index];
-        return <div key={key} style={style}>
-            <Checkbox
-                forceState={meta.selected}
-                className="ride-card-container"
-                html={<div><b>{meta.TaskId}</b><br></br>{new Date(meta.Created_Date).toLocaleDateString()}</div>}
-                onClick={(isChecked) => {
-                    onClick(meta, index, isChecked)
-                }}/>
-        </div>
-    }
+const Cards: FC<CardsProps> = ({ showMetas, onClick, height }) => {
+	const renderRow: ListRowRenderer = ({ index, key, style }): ReactNode => {
+		const meta = showMetas[index];
+		return <div key={key} style={style}>
+			{/*<ToggleButton value={meta.selected} onClick={(isChecked) => {*/}
+			{/*	onClick(meta, index, isChecked);*/}
+			{/*}}/>*/}
+		
+			{/*<ListItem onSelect={(isSelected) => onClick(<meta>index,isSelected)} />*/}
+			
+			<Checkbox
+				forceState={meta.selected}
+				// value={meta.selected} name={"hej"}
+				className='ride-card-container'
+				html={<div><b>{meta.TaskId}</b><br></br>{new Date(meta.Created_Date).toLocaleDateString()}</div>}
+				onClick={(isChecked) => {
+					onClick(meta, index, isChecked);
+				}} />
+		</div>;
+	};
 
-    // @ts-ignore
-    return <List
-         width={170}
-         height={2500}
-         rowHeight={61}
-         rowRenderer={renderRow}
-         rowCount={showMetas.length}/>
-}
+	// @ts-ignore
+	return <List
+		width={170}
+		height={height - 230}
+		rowHeight={75}
+		rowRenderer={renderRow}
+		rowCount={showMetas.length} />;
+};
 
 interface SelectMeta extends RideMeta {
-    selected: boolean;
+	selected: boolean;
 }
 
-const RideCards: FC = () => {
+const RideCards: FC<{height: number}> = ({height}) => {
+	const { metas, setSelectedMetas, showMetas, setShowMetas } = useMetasCtx();
 
-    const {metas, selectedMetas, setSelectedMetas} = useMetasCtx();
+	useEffect(() => {
+		setShowMetas(metas.map(m => ({ ...m, selected: false })));
+	}, [metas, setShowMetas]);
 
-    const [showMetas, setShowMetas] = useState<SelectMeta[]>([])
+	const onClick = (md: SelectMeta, i: number, isChecked: boolean) => {
+		const temp = [...showMetas];
+		temp[i].selected = isChecked;
+		setShowMetas(temp);
 
-    useEffect(() => {
-        // console.log("Metas:\n");
-        // console.log(metas); // ProdURL
-        setShowMetas(metas.map(m => ({...m, selected: false})))
-    }, [metas])
+		return isChecked
+			? setSelectedMetas(prev => [...prev, md])
+			: setSelectedMetas(prev => prev.filter(({ TripId }) => md.TripId !== TripId));
+	};
 
-    const onChange = ({search, startDate, endDate, reversed}: TripsOptions) => {
-        const temp: SelectMeta[] = metas
-            .filter((meta: RideMeta) => {
-                const inSearch = search === "" || meta.TaskId.toString().includes(search)
-                const date = new Date(meta.Created_Date).getTime()
-                const inDate = date >= startDate.getTime() && date <= endDate.getTime()
-                return inSearch && inDate
-            })
-            .map((meta: RideMeta) => {
-                const selected = selectedMetas.find(({TripId}) => meta.TripId === TripId) !== undefined
-                return {...meta, selected}
-            })
-        setShowMetas(reversed ? temp.reverse() : temp)
-    }
-
-    const onClick = (md: SelectMeta, i: number, isChecked: boolean) => {
-        const temp = [...showMetas]
-        temp[i].selected = isChecked
-        setShowMetas(temp)
-
-        return isChecked
-            ? setSelectedMetas(prev => [...prev, md])
-            : setSelectedMetas(prev => prev.filter(({TripId}) => md.TripId !== TripId))
-    }
-
-
-    return (
-        <div className="ride-list" style={{ position: 'absolute', height: '100vh', zIndex: 500 }}>
-            {/*<OptionsSelector onChange={onChange}/>*/}
-            <Cards showMetas={showMetas} onClick={onClick}/>
-        </div>
-    )
-}
+	return (
+		<div className='ride-list' style={{ position: 'absolute', height: '100vh', zIndex: 500 }}>
+			<Cards showMetas={showMetas} onClick={onClick} height={height} />
+		</div>
+	);
+};
 
 export default RideCards;
