--- client/src/components/RoadMeasurements/RideDetailsA.tsx
@@ -1,27 +1,0 @@
-import React, { FC } from "react";
-import MetaData from "./MetaData";
-
-import { useMetasCtx } from "../../context/MetasContext";
-import { RideMeta } from "../../models/models";
-import { Stack } from "@mui/material";
-
-const RideDetails: FC<{height: number}> = ({height}) => {
-  const { selectedMetas } = useMetasCtx();
-
-  return (
-    <Stack
-      className="meta-data"
-      style={{
-          height: height - 230,
-          overflowX: 'hidden',
-          zIndex: 500,
-      }}
-    >
-      {selectedMetas.map((meta: RideMeta, _: number) => (
-        <MetaData md={meta} key={`md-${Math.random()}`} />
-      ))}
-    </Stack>
-  );
-};
-
-export default RideDetails;
